Corrected matrix of rotation
Adjusted type for floating number arithmetic

+++ src/tools.cpp
@@ -157,21 +157,21 @@ namespace ipt {
             qz = (m10 - m01) / S;
         }
         else if ((m00 > m11) & (m00 > m22)) {
-            float S = sqrt(1.0 + m00 - m11 - m22) * 2; // S=4*qx 
+            double S = sqrt(1.0 + m00 - m11 - m22) * 2; // S=4*qx 
             qw = (m21 - m12) / S;
             qx = 0.25 * S;
             qy = (m01 + m10) / S;
             qz = (m02 + m20) / S;
         }
         else if (m11 > m22) {
-            float S = sqrt(1.0 + m11 - m00 - m22) * 2; // S=4*qy
+            double S = sqrt(1.0 + m11 - m00 - m22) * 2; // S=4*qy
             qw = (m02 - m20) / S;
             qx = (m01 + m10) / S;
             qy = 0.25 * S;
             qz = (m12 + m21) / S;
         }
         else {
-            float S = sqrt(1.0 + m22 - m00 - m11) * 2; // S=4*qz
+            double S = sqrt(1.0 + m22 - m00 - m11) * 2; // S=4*qz
             qw = (m10 - m01) / S;
             qx = (m02 + m20) / S;
             qy = (m12 + m21) / S;
